+++ src/psync/pfscrypto.c
@@ -934,9 +934,9 @@ static int psync_fs_crypto_do_finalize_log(psync_openfile_t *of, int fullsync) {
   psync_binhex(fileidhex, &fileid, sizeof(psync_fsfileid_t));
   fileidhex[sizeof(psync_fsfileid_t)]='l';
   fileidhex[sizeof(psync_fsfileid_t)+1]=0;
-  olog=psync_strcat(cachepath, PSYNC_DIRECTORY_SEPARATOR, fileidhex, NULL);
+  olog=psync_strcat(cachepath, "/", fileidhex, NULL);
   fileidhex[sizeof(psync_fsfileid_t)]='f';
-  flog=psync_strcat(cachepath, PSYNC_DIRECTORY_SEPARATOR, fileidhex, NULL);
+  flog=psync_strcat(cachepath, "/", fileidhex, NULL);
   if (unlikely_log(psync_file_rename_overwrite(olog, flog)) ||
       unlikely_log((of->logfile=psync_file_open(olog, P_O_RDWR, P_O_CREAT|P_O_TRUNC))==INVALID_HANDLE_VALUE) ||
       unlikely_log(psync_fs_crypto_init_log(of))) {
@@ -1000,7 +1000,7 @@ PSYNC_NOINLINE static void psync_fs_crypto_reset_log_to_off(psync_openfile_t *of
     psync_binhex(fileidhex, &fileid, sizeof(psync_fsfileid_t));
     fileidhex[sizeof(psync_fsfileid_t)]='l';
     fileidhex[sizeof(psync_fsfileid_t)+1]=0;
-    log=psync_strcat(cachepath, PSYNC_DIRECTORY_SEPARATOR, fileidhex, NULL);
+    log=psync_strcat(cachepath, "/", fileidhex, NULL);
     if (psync_file_delete(log))
       log_info("could not delete old log file %s", log);
     of->logfile=psync_file_open(log, P_O_RDWR, P_O_CREAT|P_O_TRUNC);
@@ -1815,7 +1815,7 @@ static void psync_fs_crypto_check_file(void *ptr, psync_pstat_fast *st) {
     return;
   ch=st->name[len-1];
   if (ch=='l' || ch=='f') {
-    path=psync_strcat((const char *)ptr, PSYNC_DIRECTORY_SEPARATOR, st->name, NULL);
+    path=psync_strcat((const char *)ptr, "/", st->name, NULL);
     psync_fs_crypto_check_log(path, st->name);
     psync_free(path);
   }